Refactor class and namespace names to use the Models namespace
