+++ tests/EntityTest.php
@@ -1,10 +1,10 @@
 <?php
 
 use Kiwilan\Ebook\Ebook;
-use Kiwilan\Ebook\Tools\BookAuthor;
-use Kiwilan\Ebook\Tools\BookContributor;
-use Kiwilan\Ebook\Tools\BookIdentifier;
-use Kiwilan\Ebook\Tools\BookMeta;
+use Kiwilan\Ebook\Models\BookAuthor;
+use Kiwilan\Ebook\Models\BookContributor;
+use Kiwilan\Ebook\Models\BookIdentifier;
+use Kiwilan\Ebook\Models\BookMeta;
 
 it('can use EbookEntity', function () {
     $ebook = Ebook::read(EPUB);
